kbs: Rename attest to attestation
Just making things a little clearer.
